chore: move scratch out of src folder
