+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
 	globals: {
 		'ts-jest': {
-			tsConfig: 'tsconfig.jest.json',
+			tsconfig: 'tsconfig.jest.json',
 			diagnostics: false
 		}
 	},
