+++ package.json
@@ -101,13 +101,11 @@
     "/LICENSE"
   ],
   "dependencies": {
-    "koa": "^2.13.1",
     "request": "^2.88.2",
     "request-promise-native": "^1.0.9",
     "uuid": "^8.3.2",
     "ws": "^7.4.2",
-    "xml2js": "^0.4.23",
-    "yargs": "^16.2.0"
+    "xml2js": "^0.4.23"
   },
   "devDependencies": {
     "@types/jest": "^26.0.20",
@@ -125,6 +123,7 @@
     "jest": "^26.6.3",
     "jest-haste-map": "^26.6.2",
     "jest-resolve": "^26.6.2",
+    "koa": "^2.13.1",
     "koa-bodyparser": "^4.3.0",
     "node-license-validator": "^1.3.0",
     "npm-scripts-info": "^0.3.9",
@@ -136,7 +135,8 @@
     "tslint": "^5.20.0",
     "tslint-config-standard": "^8.0.1",
     "typedoc": "^0.20.16",
-    "typescript": "~4.0"
+    "typescript": "~4.0",
+    "yargs": "^16.2.0"
   },
   "standard-version": {
     "message": "chore(release): %s [skip ci]",
