+++ scratch/experiments/cli_sm_gfx.js
