+++ scratch/experiments/delete.js
