+++ scratch/experiments/ping.js
