+++ scratch/experiments/read_masters.js
