+++ scratch/experiments/read_template.js
