+++ scratch/experiments/same.js
