+++ scratch/experiments/socket.js
