+++ scratch/pepuri.ts
@@ -1,24 +1,26 @@
-import { startPepTalk } from '../peptalk'
+import { startPepTalk } from '../src/peptalk'
 import * as yargs from 'yargs'
 
 let args = yargs
 	.string('host')
 	.number('port')
 	.string('_')
+	.alias('_', 'path')
+	.coerce('path', a => a.reduce(
+		(x: string, y: string) => `${x}${x.length === 0 || x.endsWith('/') ? '' : '/'}${y}`, ''))
 	.default('host', 'localhost')
 	.default('port', 8595)
-	.demandCommand(2, 3)
 	.argv
 
 console.dir(args)
 
 async function run () {
 	let pt = startPepTalk(args.host, args.port)
-	let connected = await pt.connect(true)
+	let connected = await pt.connect()
 	console.log(connected)
 	try {
-		console.log(await pt.set(args._[0], args._[1], args._[2]))
-	} catch (err) { console.error('!!!', err) }
+		console.log(await pt.uri(args.path, 'hierarchy_collection'))
+	} catch (err) { console.dir(err) }
 	await pt.close()
 }
