+++ scratch/sniff.js
