+++ tsconfig.json
@@ -9,7 +9,7 @@
   "extends": "./config/tsconfig.base",
   "include": [
     "./src/**/*"
-  ],
+, "scratch"  ],
   "exclude": [
     "node_modules/**",
     "src/**/*spec.ts",
